+++ static/js/nzbhydra.js
@@ -1853,7 +1853,7 @@ function SearchHistoryController($scope, $state, StatsService, history) {
         if (request.type == "tv") {
             stateParams.mode = "tvsearch"
         } else if (request.type == "tv") {
-            stateParams.mode = "moviesearch"
+            stateParams.mode = "movie"
         } else {
             stateParams.mode = "search"
         }
@@ -2046,9 +2046,8 @@ function SearchController($scope, $http, $stateParams, $state, SearchService, fo
     $scope.goToSearchUrl = function () {
         var stateParams = {};
         if ($scope.category.name.indexOf("movies") > -1) {
-            stateParams.mode = "moviesearch";
             stateParams.title = $scope.title;
-            stateParams.mode = "moviesearch";
+            stateParams.mode = "movie";
         } else if ($scope.category.name.indexOf("tv") > -1) {
             stateParams.mode = "tvsearch";
             stateParams.title = $scope.title;